Add files via upload
this is a very basic example of a calendar app. all we would have to do is make a generic list of events and a button that would then pick one of them and add it to the calendar.  We could make this short so it only shows a week instead of all the months out of the year but its a start from the library i was using